+++ core/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.39.0 [unreleased]
+
+- Move `ConnectionId` to `libp2p-swarm`. See [PR 3221].
+- Move `PendingPoint` to `libp2p-swarm` and make it crate-private. See [PR 3221].
+
+[PR 3221]: https://github.com/libp2p/rust-libp2p/pull/3221
+
 # 0.38.0
 
 - Remove deprecated functions `StreamMuxerExt::next_{inbound,outbound}`. See [PR 3031].
@@ -18,7 +25,7 @@
 
 - Remove `EitherUpgrade` in favor of implementing `UpgradeInfo`, `InboundUpgrade` and `OutboundUpgrade` on `either::Either`. See [PR 3339].
 
-- Remove `EitherFuture2` in favor of `EitherFuture`. See [PR XXXX].
+- Remove `EitherFuture2` in favor of `EitherFuture`. See [PR 3340].
 
 [PR 3031]: https://github.com/libp2p/rust-libp2p/pull/3031
 [PR 3058]: https://github.com/libp2p/rust-libp2p/pull/3058
@@ -28,7 +35,7 @@
 [PR 3337]: https://github.com/libp2p/rust-libp2p/pull/3337
 [PR 3338]: https://github.com/libp2p/rust-libp2p/pull/3338
 [PR 3339]: https://github.com/libp2p/rust-libp2p/pull/3339
-[PR XXXX]: https://github.com/libp2p/rust-libp2p/pull/XXXX
+[PR 3340]: https://github.com/libp2p/rust-libp2p/pull/3340
 
 # 0.37.0